+++ source/extensions/isaacsim.core.simulation_manager/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [1.1.0] - 2025-05-30
+### Changed
+- Added support for in memory stages
+
 ## [1.0.1] - 2025-05-19
 ### Changed
 - Update copyright and license to apache v2.0